autogen: refactor common parts into separate modules
